Image Detection
New code to find all the different objects in the image and put them in
a list. Each object can then be checked by using a filter. The starting
code for the filter has been implemented, but they objects currently
always pass. Specific filter code still has to be implemented (eg: area,
roundness). Also implemented code that would take all the objects that
pass the filter and put them back on the canvas with a different
grayvalue for each object.