Tweaks
Updated the algorithm. Cleaned the code. Updated GUI so its easier to
see what's happening. Saving each step in an image, so its easier to
review after all the steps are done. Changed Opening so it wont remove
the border anymore. (Well, actually added a feature to re-draw the
missing border, based on neighbour pixels.)